+++ storehaus-hbase/src/main/scala/com/twitter/storehaus/asynchbase/AsyncHBaseLongStore.scala
@@ -17,7 +17,7 @@
 package com.twitter.storehaus.asynchbase
 
 import com.twitter.storehaus.Store
-import com.twitter.util.Future
+import com.twitter.util.{Future, Time}
 import com.twitter.bijection.Injection._
 import org.hbase.async.HBaseClient
 
@@ -71,8 +71,6 @@ class AsyncHBaseLongStore(protected val quorumNames: Seq[String],
   /** Close this store and release any resources.
     * It is undefined what happens on get/multiGet after close
     */
-  override def close {
-    client.shutdown()
-  }
+  override def close(t: Time) = futurePool { client.shutdown() }
 }